Move eval example to evaluation_and_profiling
