+++ examples/evaluation_and_profiling/simple_web_query_eval/pyproject.toml
@@ -41,3 +41,6 @@ dependencies = [
 [tool.uv.sources]
 nvidia-nat = { path = "../../..", editable = true }
 nat_simple_web_query = { path = "../../getting_started/simple_web_query", editable = true }
+
+[project.entry-points.'nat.components']
+nat_simple_web_query_eval = "nat_simple_web_query_eval.register"
