+++ examples/evaluation_and_profiling/simple_web_query_eval/tests/test_atif_only_evaluator_register.py
@@ -18,9 +18,9 @@ from unittest.mock import MagicMock
 from nat.data_models.atif import ATIFAgentConfig
 from nat.data_models.atif import ATIFTrajectory
 from nat.plugins.eval.evaluator.atif_evaluator import AtifEvalSample
-from nat_simple_web_query.atif_only_evaluator_register import AtifCosineSimilarityEvaluator
-from nat_simple_web_query.atif_only_evaluator_register import AtifCosineSimilarityEvaluatorConfig
-from nat_simple_web_query.atif_only_evaluator_register import register_atif_cosine_similarity_evaluator
+from nat_simple_web_query_eval.atif_only_evaluator_register import AtifCosineSimilarityEvaluator
+from nat_simple_web_query_eval.atif_only_evaluator_register import AtifCosineSimilarityEvaluatorConfig
+from nat_simple_web_query_eval.atif_only_evaluator_register import register_atif_cosine_similarity_evaluator
 
 
 async def test_register_atif_cosine_similarity_evaluator_exposes_only_atif_lane():
